+++ app/src/main/java/foi/air/szokpt/views/app/DashboardView.kt
@@ -35,6 +35,7 @@ import foi.air.szokpt.models.CardBrandInformation
 import foi.air.szokpt.ui.components.TileSegment
 import foi.air.szokpt.ui.components.dashboard_components.BarComponent
 import foi.air.szokpt.ui.components.dashboard_components.ChartWithLegend
+import foi.air.szokpt.ui.components.dashboard_components.CustomCircularProgressBar
 import foi.air.szokpt.ui.theme.Alternative
 import foi.air.szokpt.ui.theme.BGLevelOne
 import foi.air.szokpt.ui.theme.Primary
@@ -239,43 +240,6 @@ fun TextRow(label: String, value: String, color: Color) {
         )
     }
 }
-
-
-@Composable
-fun CustomCircularProgressBar(
-    progress: Float,
-    modifier: Modifier = Modifier,
-    backgroundColor: Color = Color.Gray,
-    progressColor: Color = Color.White,
-    strokeWidth: Dp = 8.dp,
-) {
-    Box(
-        contentAlignment = Alignment.Center,
-        modifier = modifier
-            .size(110.dp)
-    ) {
-        CircularProgressIndicator(
-            progress = {
-                1f
-            },
-            modifier = Modifier.fillMaxSize(),
-            color = backgroundColor,
-            strokeWidth = strokeWidth,
-            strokeCap = StrokeCap.Round,
-        )
-
-        CircularProgressIndicator(
-            progress = {
-                progress
-            },
-            modifier = Modifier.fillMaxSize(),
-            color = progressColor,
-            strokeWidth = strokeWidth,
-            strokeCap = StrokeCap.Round,
-        )
-    }
-}
-
 
 @Composable
 fun TransactionsListTile() {
